New Files and New Approach
Added, among other things, fabric.js, which should allow the whiteboard to do exactly as the client intends. This came about due to a limitation with the basic html canvas where text could not be edited on said canvas. Fabric.js should be the key to solving that and a plethora o other issues, but as a result will require another week to investigate.