Eraser and Custom Save Files
There is an interesting quirk, where upon saving the canvas as a png file, the png might show up with a black background... Not quite sure what to do with that, but it's a thing that exists, so be careful drawing in white. As a result, the default pen color has been switched to blue just as a safety precaution for users.