Roomeet is Live!
The MVP has been built into an actual working .exe file for Windows. Unfortunately, as of this commit I cannot build it for non-Windows platforms as I do not have the tools to test on said platforms.

A key distinction going forward:
Roomeet.py is the official development code that will be built for release, and Roomeet_testbed.py is the development "playground" where new features will constantly be added and tested prior to the build release.