New Approach
Version 0.2.0 of Kivy is now available, as is the basic version of the ReadMe file. Roomeet v0.2.0 is using the new testbed file, which is basically an improved version of the original (with a canvas that can't draw over the buttons), so it's still missing a fair bit of the original functionality.